Removed Threading code from DataCollectionTask
